sync3

+++ opdr7.py
@@ -4,7 +4,7 @@ wh=input("")
 ws,hs=wh.split(",")
 print(f"calculating with and height of width {ws} and height {hs}")
 
-w=int(ws.strip())
-h=int(hs.strip())
+w=float(ws.strip())
+h=float(hs.strip())
 
 print(f"volume is {w*h/2}")
